fix babel plugin: move to main file
